Set the Github App ID via envvars
This commit sets the Derek Github Application ID via the `application`
envvar. This ID is the one associated with the registered application,
while the installation is associated the instance of Derek associated
with your Github account.

Failing to set the correct id (ie the application ID) was causing the
generation of JWT tokens to fail with an invalid permissions error.

Signed-off-by: John McCabe <john@johnmccabe.net>