Add commit message linting feature
Adds feature to .DEREK.yml `commit_linting`, which when enabled
lints commit messages with a style consistent with the guidelines
written up by Chris Beams [1]. The linting will place a label on a PR
and send a comment when it fails linting, but this is not blocking.

The label used is: review/commit-message and that will be removed
if and when the linting is fixed.

Unit test coverage is added for the linting.

The DCO checking feature was also tested and an optimization was
added so that the PR labels are only fetched once and then re-used
this is to save on API rate-limiting and server trips.

[1] https://chris.beams.io/posts/git-commit/

Signed-off-by: Alex Ellis <alexellis2@gmail.com>