Add comma separated multi labels facility
Currently to add multiple labels the maintainer must add one comment for each label.
This change enables multiple labels to be applied through  single comment, by allowing
for a comma separated list to be provided.

As the deletion of multiple labels is currently an iterative process, due to how the
github API is presented, an error could occur mid-flow causing subsequent labels to
not be applied.  This is considered of minimal concern as the results are verifiable
visually.

Signed-off-by: rgee0 <richard@technologee.co.uk>