Add PR branch validation
This  checks  if  the  branch  a  pull  request  is  raised  from  is
non-master.  Validates,  as well,  if a pull request is raised against
the  default  branch.  If the validation  is not successful,  the pull
request labels are applied: "review/target-branch" if the target
branch is not  the  default  and  "review/source-branch"
if the source branch is named  "master".  The source branch validation
leaves a warning comment,  suggesting to raise a new pull request from
a named branch

Signed-off-by: Ivana Yovcheva (VMware) <iyovcheva@vmware.com>