Add "derek-*-key" to gitignore & amend GET.md
The names of the secrets were changed following review so that they function on both
Docker and k8s.  This updates the .gitignore file to change derek_*_key to derek-*-key
for consistency.

Changed reference of "_" to "-" in GET.md and added example commands to add secrets
using Docker and k8s.

Signed-off-by: rgee0 <richard@technologee.co.uk>