Accommodate custom .CUSTOMERS via ENV var
The existing location of the .CUSTOMERS file is hard-coded which means
that anyone self hosting will need to either add an entry into the project
.CUSTOMERS, or alter the source code to point at their own location.

This change enables an over-ride of the default .CUSTOMERS location by
setting an `customers_url` environment variable.  The location should
be a `https` end point.

Test added for the new function which attempts to read the ENV var and
GET.md updated to include `customers_url` in the stack file section.

Signed-off-by: rgee0 <richard@technologee.co.uk>