Support the loading of config files via URL redirection
Expose the .DEREK.yaml field "redirect" in the config which
allows a config to be read from another location.

If the "redirect" field is non-empty the current loaded
config is ignored and the file at the "redirect" URL is treated
as the actual config.

This change is backwards compatible as the "redirect" field is
optional.

Signed-off-by: Lubomir I. Ivanov (VMware) <neolit123@gmail.com>