Accommodate OpenFaaS secrets location change
Recent changes to OpenFaaS changed the storage location of secrets.
The current Derek implementation expects to find secrets in the (now) legacy location.
This change adds functionality to default to the new location and fall back to legacy
if the new location doesnt exist so that Derek will continue to operate either side of
OpenFaaS 0.8.2

Signed-off-by: rgee0 <richard@technologee.co.uk>