Add comma separated multi labels facility
Currently to add multiple labels a maintainer must add one comment for each label.
This change enables multiple labels to be applied through a single comment, by allowing
for a comma separated list to be provided.

As the deletion of multiple labels is currently an iterative process, due to how the github API is presented, an error could occur mid-flow when deleting and cause subsequent labels to not be applied. Mitigation against this is that labels are visually verifiable.

Signed-off-by: Richard Gee <richard@technologee.co.uk>