Add support for private repositories
In order to support fetching files from repositories that are private
a new function was added called GetPrivateRepoConfig. This function
uses GitHub's API to download the file instead of download it from
the CDN.
I thought it would be best to introduce a new function instead of
rewriting the current one and changing its signature. This is because
I wanted to make sure I wasn't breaking any existing functionality.

Signed-off-by: Matias Pan <matias.pan26@gmail.com>