Add Hacktoberfest feature
This change adds the new feature hacktoberfest
The goal of this feature is to find flyby pull requests
that are considered spam.
A pull request that is a first time contributor, not signed,
and only changing markdown (md) files is assumed to be spam.
If the contributor has read the contribution guide, they
would have signed their commit, and the handler would have
no affect.
Otherwise, the pull request is closed, and an invalid
label is added indicating that this PR should not be counted

Signed-off-by: Burton Rheutan <rheutan7@gmail.com>