Force local headings to over-ride remote
A recent issue (https://github.com/alexellis/go-execute/issues/10)
indirectly highlighted that the merging of local and remote configs
was producing a superset which included duplicate headings.  The
intended mode of operation for the Issue template headings was
for the local to over-ride the remote.  This is diffent to the other
fields, such as maintainers.

This change adds the test to highlight the issue and addresses the
issue in MergeDerekRepoConfigs.

Signed-off-by: Richard Gee <richard@technologee.co.uk>